+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/model/RoomSummaryEntity.kt
@@ -21,7 +21,9 @@ import io.realm.RealmObject
 import io.realm.annotations.Index
 import io.realm.annotations.PrimaryKey
 import org.matrix.android.sdk.api.crypto.RoomEncryptionTrustLevel
+import org.matrix.android.sdk.api.extensions.tryOrNull
 import org.matrix.android.sdk.api.session.room.model.Membership
+import org.matrix.android.sdk.api.session.room.model.RoomJoinRules
 import org.matrix.android.sdk.api.session.room.model.RoomSummary
 import org.matrix.android.sdk.api.session.room.model.VersioningState
 import org.matrix.android.sdk.api.session.room.model.tag.RoomTag
@@ -239,6 +241,19 @@ internal open class RoomSummaryEntity(
         set(value) {
             if (value.name != versioningStateStr) {
                 versioningStateStr = value.name
+            }
+        }
+
+    private var joinRulesStr: String? = null
+    var joinRules: RoomJoinRules?
+        get() {
+            return joinRulesStr?.let {
+                tryOrNull { RoomJoinRules.valueOf(it) }
+            }
+        }
+        set(value) {
+            if (value?.name != joinRulesStr) {
+                joinRulesStr = value?.name
             }
         }