+++ vector/src/main/java/im/vector/app/features/home/HomeActivity.kt
@@ -55,14 +55,15 @@ import im.vector.app.features.permalink.PermalinkHandler
 import im.vector.app.features.popup.DefaultVectorAlert
 import im.vector.app.features.popup.PopupAlertManager
 import im.vector.app.features.popup.VerificationVectorAlert
+import im.vector.app.features.rageshake.ReportType
 import im.vector.app.features.rageshake.VectorUncaughtExceptionHandler
 import im.vector.app.features.settings.VectorPreferences
 import im.vector.app.features.settings.VectorSettingsActivity
-import im.vector.app.features.spaces.ShareSpaceBottomSheet
 import im.vector.app.features.spaces.SpaceCreationActivity
 import im.vector.app.features.spaces.SpacePreviewActivity
 import im.vector.app.features.spaces.SpaceSettingsMenuBottomSheet
 import im.vector.app.features.spaces.invite.SpaceInviteBottomSheet
+import im.vector.app.features.spaces.share.ShareSpaceBottomSheet
 import im.vector.app.features.themes.ThemeUtils
 import im.vector.app.features.workers.signout.ServerBackupStatusViewModel
 import im.vector.app.features.workers.signout.ServerBackupStatusViewState
@@ -215,6 +216,9 @@ class HomeActivity :
                         is HomeActivitySharedAction.OpenSpaceInvite -> {
                             SpaceInviteBottomSheet.newInstance(sharedAction.spaceId)
                                     .show(supportFragmentManager, "SPACE_INVITE")
+                        }
+                        HomeActivitySharedAction.SendSpaceFeedBack -> {
+                            bugReporter.openBugReportScreen(this, ReportType.SPACE_BETA_FEEDBACK)
                         }
                     }.exhaustive
                 }
@@ -449,11 +453,11 @@ class HomeActivity :
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
             R.id.menu_home_suggestion -> {
-                bugReporter.openBugReportScreen(this, true)
+                bugReporter.openBugReportScreen(this, ReportType.SUGGESTION)
                 return true
             }
             R.id.menu_home_report_bug -> {
-                bugReporter.openBugReportScreen(this, false)
+                bugReporter.openBugReportScreen(this, ReportType.BUG_REPORT)
                 return true
             }
             R.id.menu_home_init_sync_legacy -> {