+++ vector/src/main/java/im/vector/app/features/home/room/detail/RoomDetailViewModel.kt
@@ -449,7 +449,7 @@ class RoomDetailViewModel @AssistedInject constructor(
                 widgetSessionId = widgetSessionId.substring(0, 7)
             }
             val roomId: String = room.roomId
-            val confId = roomId.substring(1, roomId.indexOf(":") - 1) + widgetSessionId.toLowerCase(VectorLocale.applicationLocale)
+            val confId = roomId.substring(1, roomId.indexOf(":") - 1) + widgetSessionId.lowercase(VectorLocale.applicationLocale)
 
             val preferredJitsiDomain = tryOrNull {
                 rawService.getElementWellknown(session.myUserId)
@@ -832,7 +832,7 @@ class RoomDetailViewModel @AssistedInject constructor(
                                     session.spaceService().getSpace(spaceId)
                                             ?.addChildren(
                                                     state.roomId,
-                                                    listOf(session.sessionParams.homeServerHost ?: ""),
+                                                    null,
                                                     null,
                                                     true
                                             )
@@ -849,7 +849,7 @@ class RoomDetailViewModel @AssistedInject constructor(
                                     session.spaceService().getSpace(slashCommandResult.spaceId)
                                             ?.addChildren(
                                                     room.roomId,
-                                                    listOf(session.sessionParams.homeServerHost ?: ""),
+                                                    null,
                                                     null,
                                                     false
                                             )