+++ vector/src/main/java/im/vector/app/features/home/room/detail/timeline/helper/MatrixItemColorProvider.kt
@@ -83,7 +83,7 @@ class MatrixItemColorProvider @Inject constructor(
         fun getColorFromUserId(userId: String?): Int {
             var hash = 0
 
-            userId?.toList()?.map { chr -> hash = (hash shl 5) - hash + chr.toInt() }
+            userId?.toList()?.map { chr -> hash = (hash shl 5) - hash + chr.code }
 
             return getUserColorByIndex(abs(hash))
         }
@@ -104,7 +104,7 @@ class MatrixItemColorProvider @Inject constructor(
 
         @ColorRes
         private fun getColorFromRoomId(roomId: String?): Int {
-            return when ((roomId?.toList()?.sumBy { it.toInt() } ?: 0) % 3) {
+            return when ((roomId?.toList()?.sumOf { it.code } ?: 0) % 3) {
                 1    -> R.color.riotx_avatar_fill_2
                 2    -> R.color.riotx_avatar_fill_3
                 else -> R.color.riotx_avatar_fill_1