+++ vector/src/main/java/im/vector/app/features/roomprofile/settings/RoomSettingsViewState.kt
@@ -20,6 +20,8 @@ import android.net.Uri
 import com.airbnb.mvrx.Async
 import com.airbnb.mvrx.MvRxState
 import com.airbnb.mvrx.Uninitialized
+import im.vector.app.R
+import im.vector.app.core.resources.StringProvider
 import im.vector.app.features.roomprofile.RoomProfileArgs
 import org.matrix.android.sdk.api.session.room.model.GuestAccess
 import org.matrix.android.sdk.api.session.room.model.RoomHistoryVisibility
@@ -51,7 +53,8 @@ data class RoomSettingsViewState(
             val canChangeName: Boolean = false,
             val canChangeTopic: Boolean = false,
             val canChangeHistoryVisibility: Boolean = false,
-            val canChangeJoinRule: Boolean = false
+            val canChangeJoinRule: Boolean = false,
+            val canAddChildren: Boolean = false
     )
 
     sealed class AvatarAction {
@@ -66,5 +69,25 @@ data class RoomSettingsViewState(
             val newGuestAccess: GuestAccess? = null
     ) {
         fun hasChanged() = newJoinRules != null || newGuestAccess != null
+    }
+
+    fun getJoinRuleWording(stringProvider: StringProvider): String {
+        return when (val joinRule = newRoomJoinRules.newJoinRules ?: currentRoomJoinRules) {
+            RoomJoinRules.INVITE -> {
+                stringProvider.getString(R.string.room_settings_room_access_private_title)
+            }
+            RoomJoinRules.PUBLIC -> {
+                stringProvider.getString(R.string.room_settings_room_access_public_title)
+            }
+            RoomJoinRules.KNOCK -> {
+                stringProvider.getString(R.string.room_settings_room_access_entry_knock)
+            }
+            RoomJoinRules.RESTRICTED -> {
+                stringProvider.getString(R.string.room_settings_room_access_restricted_title)
+            }
+            else                     -> {
+                stringProvider.getString(R.string.room_settings_room_access_entry_unknown, joinRule.value)
+            }
+        }
     }
 }